requirements.txt
snowflake-snowpark-python[pandas]
scikit-learn
joblib
python-dotenv
